home page re=ordering
Combined the expanding and our focus sections to be side by side.  Set
all of the column sizings to xs so that these two sections will remain
side by side on all screens (still need to do some fine tuning to the
icons for this to work)

Moved the sponsors & partners sections to the bottom of the page (and
added the partners part).  Set all of the column sizing to xs so this
section will not stack on a small screen.

Added Discover as Premier Sponsor and FinTech Sandbox as Community
Partner.  Used press release wording for modals.